Break out youtube video util and actions into files organized by topic
