+++ src/actions/youtube_video_actions/index.js
@@ -1,0 +1,6 @@
+export * from './channel_actions.js';
+export * from './recommended_actions.js';
+export * from './search_video_actions.js';
+export * from './subscription_actions.js';
+export * from './trending_actions.js';
+export * from './video_player_actions.js';
