+++ src/reducers/channels_reducer.js
@@ -1,7 +1,7 @@
 import { merge } from 'lodash';
 import {
   RECEIVE_CHANNEL_INFO,
-  RECEIVE_CHANNEL_VIDEOS } from "../actions/youtube_video_actions.js";
+  RECEIVE_CHANNEL_VIDEOS } from "../actions/youtube_video_actions";
 
 let _defaultState = {};
