+++ src/util/youtube_video_util/recommended_util.js
@@ -1,0 +1,10 @@
+import * as YoutubeApi from '../youtube_api';
+
+export const fetchRecommendedVideos = () => {
+  let params = {
+    home: true,
+    access_token: localStorage.getItem('google-access-token')
+  };
+
+  return YoutubeApi.activities(params);
+};
